chore: Rename index to mod
